Make 'ShortcutSelectionLogic' Injectable.
This adds a module to provide the shortcut per-app limit
and the existing DeviceConfig flag used to control whether
that limit is applied.

Once fully integrated, tests should be updated to simply
increase the limit by injecting a different value, instead
of altering the device state by modfying DeviceConfig
values.

To complete, inject into the caller of  ChooserListAdapter.addServiceResults,
moving usage injecting this component to the caller to use.

Bug: 300157408
Test: (this change does not yet affect live code)
Flag: n/a
Change-Id: I63a661ca304c8d5653992761f10983b5776d50dd