PayloadToggle domain layer preview data population
This change introduces the logic by which Shareousel is initialized with
previews, as well as how additional previews can be loaded if requested
by the UI.

Initially, previews are taken from the EXTRA_STREAM contents of the
sharing intent. Simultaneously, a connection to the
sharing-application's cursor is established and then queried; the
initial data loaded from the cursor then replaces the initial content.

When the UI requests, additional data from the cursor is loaded into
memory; the loaded data is always contiguous, and so requests can only
be made to load the data immediately before or after whatever is already
loaded.

Once an end of the cursor has been reached (either first or last row),
any elements in the intitial selection set (EXTRA_STREAM) that have not
yet appeared in the Cursor are appended, *if* those elements would have
appeared in that direction relative to the "focused index" of the set.

In order to limit memory usage, we limit the amount of data cached in
memory from the Cursor; as additional data is loaded in one direction,
data from the other direction can be evicted from the cache.

Bug: 302691505
Flag: ACONFIG android.service.chooser.chooser_payload_toggling DEVELOPMENT
Test: atest IntentResolver-tests-unit
Change-Id: I0ba7bffc10b4369d55a14aec626ca2c22f95dbb7