Remove caching from UserScopedService
Caching the intermediate values isn't necessary and could present
an opportunity for a memory leak. Retrieved service instances should
be stored for as long as they are needed, and once out of scope,
be available for collection.

This aligns with the copy now in SystemUI: ag/26600464

Flag: None; no usages yet
Test: NA; this class is stateless
Bug: 327613051
Change-Id: I2b1225cad3d17baf1a7c45e6b5e8e9ca0a79df03