Private profile share policy delegates to parent profile
This change enables the Private tab sharing to respect
cross profile share policy and present an appropriate
message when sharing is not available.

When considering private profile as a source or target,
the parent profile is substituded when checking whether
any intents can be forwarded.

Specialized subclasses of DevicePolicyBlockerEmptyState
are eliminated in favor of creating instances as needed
and returned by NoCrossProfileEmptyStateProvider. They
are now initialized directly with the required content
instead of resource Ids.

Resource are moved to DevicePolicyResources where they
are shared with ResolverActivity.

When sharing content to Private from Work when blocked
by policy the message will now correctly refer to
'Private apps'.

Test: atest NoAppsAvailableEmptyStateProviderTest
Test: manual, install TestDPC, setup private space;
      test verious permutations
Flag: EXEMPT bugfix
Bug: 324428064
Bug: 300157408
Change-Id: Ieb8e725191691ea92f2994a693086c2029452365