Payload selection change UI improvements.
The main idea is to postpone the new pager adapter (along with the list
adapters) attachment to the view until the data is fully loaded thus
minimizing the number of UI updates.
The old adapter gets destroyed but kept attached to the view and the
destroyed status of the list adapters is used to derived the appropriate
UI state (make target views disabled).
Profile tab switchng is disabled for an active (executing or pending)
selection callback. This is due to synchronization challenges between
the old (attached) and the new (detached) pager adapters, and difficulty
of tracking loading state for multiple adapters in the legacy code.
Known issues:
 * profile buttons does not have a disabled state;
 * profile buttons use drawable with a fade-in animation that causes a
   flashing effect;
 * target icons are animated in that causes a flashing effect;
 * UI is not disabled when launched into an active selection callback
   (i.e. when re-created on rotation);
 * there are some visual artifacts when navigating between tabs after a
   selection change has been changed.

Bug: 325465291
Test: atest IntentResolver-tests-unit
Test: atest IntentResolver-tests-activity
Test: manual testing with ShatTest app

Change-Id: Ic9caa7c229b5ec0547be43974fa0c3536ecde095