Block taking recents screenshot if private profile available
Sharesheet shows separate tab when private profile is on.
User can move sharesheet to the background (e.g. by launching
other app using shortcut) and then lock private
profile. Recents preview screen shows screenshot
of the Sharesheet at the last moment it was in foreground
(when the space was unlocked). As a result Recents preview
can leak whether private profile exists on device.
This change sets setting to show blank
screen in the recents when private profile is avaialble.

Bug: 333387696
Test: manual, https://buganizer.corp.google.com/issues/333387696#comment8
Change-Id: I8758cd0a311ce93a03d998c4ada8999a6d02246d