Fix incorrect cross-profile share check.
When this code was updated for new profiles infrastructure (ag/26444390)
a mistake was made in the interpretation of (tabOwnerUserHandleForLaunch),
mistaking it for 'tabOwner', ie: the user of the tab being presented. It
is not. This value represents the user which launched share sheet.

As a result, since this CL, no cross-profile "Blocked by IT" messages
would appear when no targets are available for cross-profile sharing.

The resulting UI would show an empty tab, or fall back to the "No Apps"
UI message.

While the UI has test coverage for this, the tests are run by forcing
calls to the reponsible component to return true or false, without
validation of the user ids provided.

Bug: 335142494
Test: atest IntentResolver-tests-unit
Flag: NONE
Change-Id: I32f455611e6ae6307d49c86d480962e3426b54c3