Update Preview and Empty State with materialColor
Colors for text and background in resolver_empty_state
as well as content_preview_file_text do not match the
colors used in chooser_grid_preview_text.xml. This
results in the mixing of DeviceDefault colors with
Material colors, which could show black-on-black or
white-on-white.
Update them to all use the OnSurface colors to be
consistent with chooser_grid_preview_text.xml.

Bug: 338107805
Test: manual - see hsv in issue

Change-Id: Id24d8c66cf2b7985ee6de4ac4549a14fceea800e