Shareousel: Maintain cursor order for shated items
Add a position property to PreviewModel class to track relative order of
items. For each item, the initial value is artificial and derived from
the order of the initially shared items and is updated upon reading the
additional items cursor.
Upon sharing, If the selection has not change, the items will be shared
in their original order; If the selection has changed, the order of the
items will be affected by the observed items order in the cursor.

Fix: 329683774
Test: manual testing
Test: atest IntentResolver-tests-unit
Test: atest IntentResolver-tests-activity
Flag: android.service.chooser.chooser_payload_toggling
Change-Id: Ie552887702cde75cb1a05ed3ec5415f4f4a5c8dc