Reclalculate drawer offset even when the targets list has not changed
Shareousel has a variable-sized preview that affects the overall drawer
offset. Therefore, we cannot early return from the offset calcualtion
based solely on the condition that the targets list has not changed.
Empirically, this introduces a minor overhead of the offset
recalculation when the drawor is snapped to/from the top.

Fix: 347316548
Test: manual testing with injected debug logging
Flag: com.android.intentresolver.fix_missing_drawer_offset_calculation
Change-Id: I69ec7a5e05fad0d6607e3614208b30910a3a17ba