Per-profile caller-provided direct target collections
Instead of (partially) relying on the active profile value when deciding
whether the caller-provided direct targets should be added to the list,
define a list of caller-provided targets for each profile and use those
lists unconditionally.

Bug: 343300158
Test: manual tests: launche the ShareTest app from different profiles
and confirm that the caller-provided targes are added only to the right
profile (the profile the Chooser was launched from).
Flag: EXEMPT refactoring

Change-Id: I9c4ba35d48a090b21f96ddf4723f7b28654acb48