CachingTargetDataLoader to cache bitmaps and not drawables
A preparation refactoring. Make ChachingTargetDataLoader to cache icon
bitmaps instead of the end drawables. Drawables carry state and the same
drawable can be used in two places in the target grid (i.e. in the
ranked targets row and in the all-target grid) yet the new hover effect
needs to be provided independently for each position.

Bug: 295175912
Test: manual basic functinality testing including Shareousel selection
change.
Test: atest IntentResolver-tests-unit
Flag: EXEMPT refactoring
Change-Id: I1c5e6333310e0984e39e22cab8cf162f2f31d6e8

Change-Id: I4a3b5dff8fcfcffe7742be8ea3bd348351332c9e