Fix Shareousel crash when recomposing
Recomposition could cause a crash when initial position was no longer in
the LazyRow. This fixes it by encapsulating the problematic array
indexing within the remember block so that it is only performed on the
first composition when it is guaranteed to be a valid index position.

Test: Manual test using ShareTest
BUG: 383852288
Flag: EXEMPT Bugfix
Change-Id: Ie6b322d449ddc5c83d9d4e481967c2511b165290