Allow text toggle for shares from sysui.
Criteria will need to be refined in a follow-up CL, but getting the
parts in place behind a flag here.

Preview UI change is minimal. Test has some duplicated boilerplate but
choosing not to refactor that because the UI itself is a better refactor
candidate but outside the scope of this CL.

Bug: 242791070
Test: atest ChooserRequestTest
Test: atest FilesPlusTExtContentPreviewUiTest
Test: manual end-to-end test with feature flag on.
Flag: com.android.systemui.shared.screenshot_context_url
Change-Id: I6eaca454a88e8b8c8a196d825c7ac3daee437109