Shareousel: add item position to item content description.
Add known item cursor position to the item content description
annotation when showing more than one item. This works well when the
data provided by the app is consistent (the expected majority of the
cases).
The position annotation will be ommited in cases when we show the
initially shared items before the cursor is loaded and for the initially
shared items that were missing in the cursor.

Fix: 379032721
Test: manual testing
Flag: com.android.intentresolver.announce_shareousel_item_list_position
Change-Id: Ie669db3897961d2cf09a30ef229e10badb501566